Topology: NHLT: Intel: SSP: Add support for blob format v3.0
This patch adds the 3.0 format that is used for the PTL
ACE3.x platform. Blob 3.0 is based on blob 1.5 but it adds
DMA control registers settings and has some changes in
registers bitfields.

This patch also fixes some compile warnings when NHLT_DEBUG is
set for comparing int with unsigned in ssp-debug.c for code for
all blobs variants.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>